Simplified the client of the Zen inline-editor by only requiring the client do a "require(['zen/inline-editor'], ... ; and by cleaning up the vestiges of the editor after its use.
TODO: Find a much better way to trigger the loading of the editor.
Now the user must pass the mouse pointer over the element
"pic-slot-4" to get the editor to load. There is a commented-out
section of inject-editor.js that shows how to create a cross-browser
keypress handler, which could be used as a better trigger. To use
that instead of the mouse trigger, it's merely necessary to copy
the code from the mouse.enter event handler to the appropriate
case in the keypress handler's switch statement. But testing is
required. I could not get several of the cases to execute.